Shoehorn some i10n into the joining page's form.
This implementation uses a singleton to hold resource strings, whose
defaults are set as the name of the message. This makes it easy to
use for the default case, and should make it simple enough to add
trasnlations everywhere else.

While this commit includes changes to the fr/joining page they are
by no means correct, and need actual translations putting in.