Allow for the team-status image being live but missing.
Previously, this would just 404, which is a bit ugly. Instead, we
now return the same image as for the case where the image isn't
live yet. While this possibly makes it a little harder to notice
and debug cases where the image _should_ be live, it gives a nicer
result on the live site.

Change-Id: If39cd60828b55bb7dbaac9d149edd41e46e42d10