MCI docs: Add link to more info on servo-endstops
There is a section in the servo programming docs that talks about how
the value sent to the servo relates to the angle of the servo.  This
section provides details about how many servos can be forced to break
their endstops if the signals provided to them are out of some
acceptable range.

This patch adds a link from the warning about this in the Manual
Control Interface docs to the section in the programming docs with
more information.  It also adds an anchor to the section heading in
the programming docs page so this can happen.