Replace missing timelapse with youtube video
This video is now on youtube, and no longer at the url previously
linked.  Drop in a youtube iframe with tweaks to restore the video.

This also modifies the CSS to allow for iframes that claim to be
videos, making them look just like real <video>s or <img>s.

Change-Id: I4e5dc94dcfb47bfc1ab70ab1933f8af42133d6ae