Fix #1643: Reduce robots access to /[c]git.
This removes access to /git/ and severly limits /cgit/. I've intentionally
not blocked the whole of /cgit/ as I'm pretty sure that having some
of it indexed is quite useful. These changes disallow access to any
commits, refs or diffs, as well as to anything on a specific branch
or at a specific revision. The result just leaves access to the main
pages for the projects: log, tree, overview; plus the main cgit page.

This commit uses wildcard patterns that Google [1] and Microsoft [2]
both claim to support, though support elsewhere may be less complete.

[1] http://support.google.com/webmasters/bin/answer.py?hl=en&answer=156449&from=35237&rd=1
[2] http://www.bing.com/blogs/site_blogs/b/webmaster/archive/2008/06/03/robots-exclusion-protocol-joining-together-to-provide-better-documentation.aspx

Change-Id: If4c81288fe5cbab40785f82bf32d8ac8f39ab5f4