Fix team blogs display and related warnings
A while back we had all the team properties being stored as
LiveStatusItem instances. Since this appeared redundant it was
(mostly) removed in 0e99c956. Unfortunately it appears that
the feed usage of it was overlooked (and the error was lost in
a barage of other warnings we got from the same area of code).

Reinstate using a LiveStatusItem for the feed if it is non-empty
and resolve all the other warnings in the same function by explicitly
constructing each object layer as we go.

Change-Id: I825a6364632346e91d68b4fd9a44437550f19ebe