Add three social media logos to bottom of page
Add three logos for Facebook, Twitter and Youtube with links to our
pages on each site. Also resized the feed logo to 25x25px as the minimum
requirement by some of the sites was larger than 14, so it was changed
to make them all look the same.
For example, the YouTube restrictions
http://www.youtube.com/yt/brand/en-GB/using-logo.html restricts us from
going below 25px high.

This is the least intrusive method of adding these links that I can
think of, and adblock still will hide them.

Change-Id: I29e24b23ce73df95f6c8e76ee610df8b8a05be8a