Fix #1737: Add a page documenting user-accounts.
This adds a page to the docs which explains how to get a user account,
what they are useful for & what to do if you forget something about it.
Also link to this page from the team-leader resources, so they can see
what we're telling the students about their accounts.

Change-Id: I6722ae16eb2f12f6a8dcc3fd595023ce6917469e