Fix #2027: Replace jointio hardware docs with Ruggeduino.
This removes most of the jointio board docs & links, with the
exception of the software pages and links that I believe are removed
via other (pending) commits.

Content credit Joe Hacker.
Note: The case dimensions are not present, but this isn't a blocker
for kickstart, while the rest of this page is.

Change-Id: I46aaeeabbe67f963faefc74a8f1a3d9f108e1b4c