Add board photo and diagram to the MCv4b docs
This also removes 'overflow: hidden' from the docs paragraph tags so
that the second paragraph of text doesn't get pushed below the board
photo. The history isn't entirely clear as to why this was here. From
what I can tell/remember, we wanted sections (delimited by headings) to
clear floated images. We originally did this by setting 'clear: both' on
the heading tags, but changed this to the 'overflow: hidden' thing for
some unknown reason. I've changed this back to the heading 'clear: both'
for the docs and checked all of the docs pages, nothing seems to have
been adversely affected.

Change-Id: I9dcbd3d21e6d7c86a2883d21dcd4929c8804e92c