Ignore cgit & trac pages.
There are a couple of links to each of these from the main site
(various board docs & the get involved pages respectively), which
caused issues when this script was run against the live site. These
were not seen in development as the target pages don't tend to exist
on dev servers. I've now run this against the live site remotely,
and it completes fairly quickly.

Change-Id: Iebc2bdc9ddb610599d110c1b74f2bf4a4b7fae2b