Fix the styling of the teams page hints.
They're intended to appear only below the teams listings, but will
appear by the side without the clear as a result of the list being
floated into a table-ish shape.
Also ensure that the gap is a reasonable size by adding styling to
the last list element (this can't be added to the paragraph, as then
it would need to butt against the ul, which is an unknown distance
above).

Change-Id: I4c832c7f37a047dc9fe092e4e00c5cb83b01a7de