Make the styling of code and non-code <pre>s common.
Much of these rules is common, and now that the non-code one is
expressed in a general manner, it doesn't need to be in the _extra
file. This change removes the duplicate code and only slightly
tweaks the existing pre code rule in a way that is overriden by the
syntax-highlighting stylesheet anyway.

Change-Id: I5b86862cab1f2b190f3862e85b85a3aa93954dc4