Add information about IDE User Settings.
Users can configure various behaviours in the IDE, but this isn't
very well advertised. This new page serves to advertise this, but
deliberately doesn't go into much detail about _what_ can be configured
as this information is present on the settings page, and duplicating
it here would only increase the maintenance burden.

2013-11-28 Patch set 2: Grammar fix ('whose').

Change-Id: I41409bd20f5cc78c27be564fbbadb5c1efb7b7f4