Fix #2136: Update the Walkthrough to match the IDE.
While the walkthrough does warn that the screenshots may not exactly
match what a user sees, people don't read these notices, instead
focussing on the images. As a result, this commit updates all of
the screenshots of the IDE to show it as it is currently, and ensures
that they all use matching code, messages, etc. Also some minor
tweaks to the text around them to match.

Change-Id: If6ae10842fa6f726d3d1ddcada4e92de2dfa97a4