Add description, keywords to the token_ring page.
The message that pyenv emits at the students when the SRIC bus is
incomplete does link to this page, but most people don't read that
far into the 4 line message! By adding a liberal sprikling of related
keywords to this page, it should appear higher in search results,
especially for 'SRIC bus', which is the most likely search from the
error screen.

Change-Id: I94448a4952a82473ad59fde71dca4a6329dcc31d