Allow overriding of code highlighting language
At the moment, all of our code blocks default to being highlighted as
if they were Python.  This is because we add the 'python' class to
them all, which highlightjs picks up and uses to perform the
highlighting.  This patch allows one to select a different language
for an individual code block by labelling it with two classes:

 1) "override-lang" to prevent it from being labelled with the
    "python" class

 2) The class appropriate for the language one wants to highlight.
    For example, "cpp" -- which is why I'm adding this feature.
    (I want to be able to include some C++ from the default Ruggeduino
    firmware)

In markdown, one has to write the <pre> and <code> tags out manually
to achieve this.  A line of "~~~~" thus expands to this:

<pre><code class="override-lang cpp">

To get some C++ highlighted.

Change-Id: I08fa928c7205201bf921b422b73e01099777f927