Add the interactive troubleshooter
This patch implements the interactive troubleshooter as per trac
ticket #1594 and #1011. A link has been added to the documentation page
to access the troubleshooter.

The troubleshooter consists of 4 main files: data.json contains the tree
of questions and answers, main.js is the main logic for the
troubleshooter, styles.css contains styles specific to the
troubleshooter and then interactive_troubleshooter connects everything
together.

This commit is a squashed commit from the troubleshooter branch of
srweb-devel to ease merging via gerrit and the work has been done by
Chris Baines, Tom Leese, Freddie Nash and Helena Weare.

Change-Id: I02b382c37041bf4ad5afccc63434ada02ddff22a