Add testing of the troubleshooter's data file.
This is written in python as I didn't feel like adding a PHP unit-
test framework, and lots of SR projects already use nose, so there's
a pretty good chance blueshirts will already have it around.

These check that the file exists, in the right place, is valid json,
and matches the shema laid out by the current state of the file.
It also checks that all the quetions (other than the root) are linked
from another answer, and that all the links to questions are to ones
which actually exist.

2014-02-19 Patch set 2: rebased.
2014-02-19 Patch set 3: support python 3 by not using iteritems().

Change-Id: Ia11318895e6548819e6ac31a814001254d5db41f