Move the /comp redirect to being a page, update for SR2014.
This is now supported, and works better as the whole structure of
the site is handled in the same way, rather than partly being in
the .htaccess file (which everyone has to modify locally, too).

The update for this year points this link at the event page, as it
did last year, though we may want to consider other alternatives.

Change-Id: I7ca573e164c707d4a633031627782be4fb925ccd