Move the microgames redirect to a file.
This is now supported, and works better as the whole structure of
the site is handled in the same way, rather than partly being in
the .htaccess file (which everyone has to modify locally, too).

It also has the benefit that it works properly on local srweb instances,
which the .htaccess redirects do not.

Change-Id: Idaa5bbd1a1e2b202800ccaaa4c719e3539c7a809