Enable rewriting of the '/comp' url
This was originally disabled so that the redirect in this .htaccess
could redirect '/comp' to the competition page.

However this redirect was moved in 454a137 to another location, but the
rewriting was still disabled, so the '/comp' url never made it to the
rest of srweb. I didn't catch this due to browser caching, presumably.