Move explanation of common errors to python troubleshooting
This moves the 'Common Errors and mistakes' appendix from the tutorial
(where it just hung off the end, and wasn't referenced) to being
the content of the python troubleshooting page. For the moment, this
is the only content so update the description, but leave a TODO comment
to show the original intent of the page.

Change-Id: I9ffdad52f83e4ecefb43867460786f2daa89f68c