Keep the previous match on the front page schedule.
Highlight the current match if there is one, and use the next match
as the key to which ones to show. This allows us to keep the previous
match on display for context. It also resolves a bug where the
schedule would show matches 0-9 if there wasn't a current match (as
was the case between periods).

Change-Id: I5ea2d1542a048e40833a7f3311aa49e5c6635fc5