Add a page showing match points.
We let the user select an arbitrary range of matches, fetch the data
and then show the scores for those matches. At the moment the team
selector doesn't do anything, but eventually it will fill the matches
box with the matches that that team was in, and also highlight the
team in the results.

Change-Id: Id62ab14f6646b2a6539d73002382c2fbb46f6c60