Extract common logic to add error handling.
We do basically the same thing with each match's data, so it might
as well only appear once. This also makes it much more sane to now
add the better handling for the case where the given match doesn't
appear to exist ('previous' before matches have started, and 'current'
during gaps and off either end of the schedule).

Change-Id: Ib9b2f3c78fc747353efad036305cc06195278046