Correct kickstart date; confirm London location
The Kickstart date has changed (see list thread "kickstart has moved").
This patch corrects the date on the homepage and on the key dates page,
as well as updating the London Kickstart location to "Facebook HQ" on the
latter. Those two pages have been tested using a local apache instance
and my browser.