Normalise case for 'Kickstart'
Since new articles will be using 'Kickstart' over 'KickStart', this
patch normalises the case of any 'KickStart's in various navigation
bars.

The reason for using 'Kickstart' is simply down to it looking more
natural, and there is approximately a 50:50 split across the website at
the moment (though that does include historic news articles).

Change-Id: I34a19d33d533d0721a5953fb4b55306223adba23