Expand the description of the SR2015 game
This replaces the one-liner on the /schools/game page with an actual
description of the game for outside audiences, who are unlikely to
want to read through the full ruleboook. It also serves as a simple
introduction for competitors.

Patch set 2: Remove implication that teams may only score from one
             flag, shrink and reposition image now we don't need to
             bodge for better line wrapping.

Change-Id: I59a99e563b1352b12ff30a26a823da59dbaa0aae