Change default .htaccess RewriteBase
Since the dawn of time the default RewriteBase for .htaccess has been
'/~chris/srweb'.

This seems wrong since it means that for most people (who are using a
badger/saffron VM situation, which, to me at least, is the 'official'
means of getting a development version of srweb) have to resolve local
change conflicts every time a commit which changes .htaccess gets pulled
in.

Whilst I'm sure there are ways of automating the process of changing the
RewriteBase line, it seems sensible to me to change it to a saner
default value.