Draw a diagram of the v4 power board
The intention is for this to illustrate where the connectors, LEDs and
other items of note are on the board, much in the same way as the
previous power boards documentation.

While I posess very few artistic skills, I this is a reasonable first
revision. Currently this is not linked to by any part of the website.

Patch Set #2: Make SVG standards-compliant (and actually render)

Change-Id: I239e7060e541b23faaff0b63cbd421de79bcab61