Add links to the microgames
Having spoken with team members who were able only attend the morning
of the (Southampton) Kickstart, they noted that it would be great
to be able to access these after the event, but that they hadn't
been able to find them. Adding specific links from the event pages
as well as a general link from the kickstart page should enable the
microgames PDFs to be found.

Change-Id: If02c67fca6017565bf9424063312ac6066bc3436