Note that servo positions should be integers
While this has actually been fixed in 'sr-robot', it's not been released
to students yet. Rather than leaving the docs as it is in the hope that
the update will be released soon, I think it's better to explicitly
update the docs to match reality.

It would be trivial to reverse this change after all.

Change-Id: Ia7fc55620d796550b33c625ea7bebb217191ba1b