Merge multiple definitions of items
While it turns out that HTML5 is ok with having multiple dd elements
in the dl, it's not very visually pleasing to have horizontal lines
between each of the parts (and doesn't really make sense). Instead
just have each defintion part flow from the others and use <br />
where neccesary to keep the line breaks (since this is much more
visible in the source than a couple of trailing spaces).

Change-Id: Ib5a213880beaee9c4fcec0fa86cccdfa33096148