Require explicit marking of code language
Remove the automatic application of 'python' to all the code blocks,
allowing other lanaguages to more easily be supported in the source.
Explicitly mark the remaining unmarked block as python, simplifying
the task of using them with other markdown parsers which increasingly
support syntax-highlighting. Replace override-lang workarounds where
they appear with the appropriate markdown syntax.

Notably this does not remove the <pre class="not-code"> blocks, which
aren't code and thus are correct to remain as just <pre> blocks.

Change-Id: I93d05b3bfc989f1e2fbff6577f6c66314c7606b9