Fix #2151: Add TLAs to the main /teams/ page
We use these TLAs a lot, and we want to encourage everyone to get
used to using them. Thus publicise them here as well. This also
makes this page more consistent with the other places we refer to
teams on the website, which almost universally prefix the name
with the TLA like this.

Change-Id: I90204a41f9a93cc1072e727ee9b957bada17846a