Fix JavaScript URL for home page video
It appears there was a very hard-to-spot bug in the original patch:
`{$root_uri}` includes a trailing slash. In the JavaScript import for
the video, an extra slash was added. If there is an Apache rewrite rule
for, say, `/srweb/`, Apache silently fixes this, meaning that it was not
found in verification. In production, Apache does not fix it, meaning
that the "Watch video" link falls back to linking to YouTube.

Also move background image into the right folder.