Publish draft pit rules
This PDF is from the 2nd revision of an ODT attached to ticket #2983.
Try saying that with a mouthful.

My intention is to not link to this on the public website, but to
instead link to it from an email to teachers. That avoids mucho
attachment wrangling.

Patch set 2: with minor fixes plus shiny headings

Change-Id: Ia08e60caed23be5bc4090fc3cc8685e28851103d