Link email addresses
It's not clear why these weren't linked, though I suspect it was
hoped that the markdown processing would do so automatically as
it does for http:// links (which it doesn't). This email address
is already linked like this from elsewhere, so this change doesn't
open it up to any additional spam.