Add meaning of world and polar coordinates
This complements the information added for orientation values in
I908ce743f24e2859f4ba793a070d5b04b4e30c4f and notes a small
inconsistency with them in the case of rot_x.

This information was determined experimentally and should be
verified.

Patch set 3: remove info box which turned out to be worrying about
             things which weren't consistent in the first place.

Change-Id: I8c648c9e4b0597d3c7b3ddc210d412581aef0016